development/backend-dev/requirements.txt
fastapi
uvicorn[standard]
httpx
pandas
ta
jinja2
openai
PyJWT
pydantic
statsmodels
scipy
google-cloud-firestore
google-cloud-storage
pyyaml
python-dotenv
python-jose[cryptography]
python-telegram-bot==21.2
telegram-webapp-auth==3.0.5
ulid-py
plotly
# Новые зависимости для архитектурных улучшений
celery[redis]
redis
prometheus-client
psutil
aiogram>=3.0.0
ccxt
websockets
